requirements.txt
requests
slack_sdk
apscheduler
django
django_rest_framework
